requirements.txt
flask
langchain
langchain-openai
chromadb
mysql-connector-python
redis
gunicorn
python-dotenv